New nodes without preset pos now next to neighbors
- When a new node is added to a model with a preset positions it will not
have a preset position of its own to begin with. This previously caused it to
initialize in the middle of the graph. Now, the addition of such a node to a
model containing a preset pos for at least some nodes causes it to initialize
at the centroid of other connected nodes.